fix(vortex_utils): move H264Decoder into a separate file
