+++ vortex_utils/README.md
@@ -1,2 +1,2 @@
-# H264Decoder
+# H264Decoder (gst_utils.py)
 Install the dependencies by running the following script: [ci_install_dependencies.sh](/scripts/ci_install_dependencies.sh)
